feat : 리그게시판 프론트 양식 일치시킴, 댓글crud, 댓글작성시 게시판댓글수 올라감, getCommentListByFreeboardId(int freeboardId) 메소드로 댓글불러올수 있음 , 거래게시판 crud, session에서 admin까지 가져옴

+++ src/main/java/com/fp/shuttlecock/leagueboard/LeagueboardController.java
@@ -17,7 +17,6 @@ import jakarta.servlet.http.HttpSession;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-
 @Controller
 public class LeagueboardController {
 	@Autowired
@@ -28,100 +27,90 @@ public class LeagueboardController {
 		System.out.println(pageRequest);
 		List<LeagueboardDTO> leagueboardList = leagueservice.getAllLeaguePostByPage(pageRequest);
 		int totalPosts = leagueservice.countLeaguePosts();
-		PageResponseDTO pageResponse = new PageResponseDTO().builder()
-				.total(totalPosts)
-				.pageAmount(pageRequest.getAmount())
-				.pageRequest(pageRequest)
-				.build();
+		PageResponseDTO pageResponse = new PageResponseDTO().builder().total(totalPosts)
+				.pageAmount(pageRequest.getAmount()).pageRequest(pageRequest).build();
 
 		model.addAttribute("leagueboardList", leagueboardList);
 		model.addAttribute("pageInfo", pageResponse);
 		return "/LeagueBoard/LeagueBoard";
 	}
-	
+
 	@GetMapping("/LeagueBoard/search")
-	public String getLeaguePostByTitleOrUserId(@RequestParam String searchKeyword, 
-			String dropdown, Model model, PageRequestDTO pageRequest) {
-		System.out.println(pageRequest);
+	public String getLeaguePostByTitleOrUserId(@RequestParam String searchKeyword, String dropdown, Model model,
+			PageRequestDTO pageRequest) {
 		List<LeagueboardDTO> leagueboardList = new ArrayList<LeagueboardDTO>();
 		pageRequest.setSearchKeyword(searchKeyword);
 		pageRequest.setCategory(dropdown);
-		System.out.println(pageRequest.getCategory());
 		leagueboardList = leagueservice.getSearchedLeaguePost(pageRequest);
 		int totalPosts = leagueservice.countSearchedLeaguePosts(pageRequest);
-		PageResponseDTO pageResponse = new PageResponseDTO().builder()
-				.total(totalPosts)
-				.pageAmount(pageRequest.getAmount())
-				.pageRequest(pageRequest)
-				.build();
-		
+		PageResponseDTO pageResponse = new PageResponseDTO().builder().total(totalPosts)
+				.pageAmount(pageRequest.getAmount()).pageRequest(pageRequest).build();
+
 		model.addAttribute("leagueboardList", leagueboardList);
 		model.addAttribute("pageInfo", pageResponse);
 		return "/LeagueBoard/LeagueBoard";
 	}
-	
+
 	@GetMapping("/LeagueBoard/{leagueboardId}")
 	public String getLeaguePost(@PathVariable int leagueboardId, Model model) {
 		LeagueboardDTO leagueboard = leagueservice.getLeaguePostById(leagueboardId);
 		model.addAttribute("leagueboard", leagueboard);
 		return "/LeagueBoard/LeagueDetail";
 	}
-	
+
 	@GetMapping("/LeagueBoard/insert")
 	public String insertLeaguePostForm() {
 		return "/LeagueBoard/LeagueRegister";
 	}
-	
+
 	@PostMapping("/LeagueBoard/insert")
-	public String insertLeaguePost(LeagueboardDTO leagueboardDTO) {
-		String user = "user1";
-		leagueboardDTO.setUser_userId(user);
+	public String insertLeaguePost(LeagueboardDTO leagueboardDTO, HttpSession session) {
+		leagueboardDTO.setUser_userId(String.valueOf(session.getAttribute("userId")));
 		leagueservice.insertLeaguePost(leagueboardDTO);
 		leagueservice.increaseWinnerPoint(leagueboardDTO.getWinner());
 		leagueservice.increaseLoserPoint(leagueboardDTO.getLoser());
+		leagueservice.increaseWriteCount(leagueboardDTO.getUser_userId());
 		return "redirect:/LeagueBoard";
 	}
-	
+
 	@GetMapping("/LeagueBoard/update/{leagueboardId}")
 	public String updateLeaguePostForm(@PathVariable int leagueboardId, Model model) throws Exception {
 		LeagueboardDTO leagueboard = leagueservice.getLeaguePostById(leagueboardId);
 		model.addAttribute("leagueboard", leagueboard);
 		return "/LeagueBoard/LeagueUpdate";
 	}
-	
+
 	@PostMapping("/LeagueBoard/update")
-	public String updateLeaguePost(LeagueboardDTO leagueboardDTO) {
-		System.out.println("리그보드 업데이트");
-		//session객체 받아오기, user를 session에서 받아온 userId로 설정하기
-		String user = "user1";
-		leagueboardDTO.setUser_userId(user);
+	public String updateLeaguePost(LeagueboardDTO leagueboardDTO, HttpSession session) {
 		boolean result = false;
-		//UserDTO user = (UserDTO) session.getAttribute("user");
-		//if (user != null && (user.getuserId() == 0 || boardService.isAuthor(newBoard.getFileNo(), user.getId()))) {
-				result = leagueservice.updateLeaguePost(leagueboardDTO);
-				if (result) {
-					return "redirect:/LeagueBoard/" + leagueboardDTO.getLeagueboardId();
+		if (String.valueOf(session.getAttribute("userId")) != null &&
+				leagueboardDTO.getUser_userId().equals(String.valueOf(session.getAttribute("userId")))) {
+			result = leagueservice.updateLeaguePost(leagueboardDTO);
+			if (result) {
+				return "redirect:/LeagueBoard/" + leagueboardDTO.getLeagueboardId();
+			} else {
+				return "error";
+			}
 		}
-		System.out.println(leagueboardDTO.getLeagueboardId());
-		return "error";
-	//}
+		return "/login";
 	}
-	
+
 	@DeleteMapping(value = "/LeagueBoard/{leagueboardId}")
-	public ResponseEntity<String> deleteLeaguePost(@PathVariable("leagueboardId") int leagueboardId, HttpSession session) {
+	public ResponseEntity<String> deleteLeaguePost(@PathVariable("leagueboardId") int leagueboardId,
+			HttpSession session) {
 		boolean result = false;
-		//UserDTO user = (UserDTO) session.getAttribute("user");
-		//if (user != null && (user.getAuthor() == 0 || boardService.isAuthor(fileNo, user.getId()))) {
-				result = leagueservice.deleteLeaguePost(leagueboardId);
-				if(result) {
-					return ResponseEntity.ok("글 삭제 성공");
-				
-				}
-				return ResponseEntity.status(500).body("글 삭제 실패");
-//			}
-//		} else {
-//			return ResponseEntity.status(403).body("삭제 권한 없음"); // 권한 없을 경우 403 에러 반환
-//		}
+		LeagueboardDTO leagueboardDTO = leagueservice.getLeaguePostById(leagueboardId);
+		if (String.valueOf(session.getAttribute("userId")) != null && ((Integer) session.getAttribute("admin") == 1
+				|| leagueboardDTO.getUser_userId().equals(String.valueOf(session.getAttribute("userId"))))) {
+			result = leagueservice.deleteLeaguePost(leagueboardId);
+			if (result) {
+				return ResponseEntity.ok("글 삭제 성공");
+
+			}
+			return ResponseEntity.status(500).body("글 삭제 실패");
+		} else {
+			return ResponseEntity.status(403).body("삭제 권한 없음");
+		}
 	}
-	
+
 }
